refactor: add atomic web design structure
