+++ src/components/atoms/Separator/styles.ts
