+++ src/components/atoms/Separator/types.ts
