+++ src/components/atoms/Text/types.ts
