+++ src/components/templates/PageLayout/styles.ts
@@ -1,10 +1,15 @@
-import { SideMenu as DefaultSideMenu } from '$/components/SideMenu';
+import { SideMenu as DefaultSideMenu } from '$/components/organisms/SideMenu';
 import { from } from '$/styles/utils/responsive';
 import styled from 'styled-components';
 
 export const Container = styled.div`
   display: flex;
+  flex-direction: column;
   height: 100vh;
+
+  ${from['tabletPortrait']} {
+    flex-direction: row;
+  }
 `;
 
 export const SideMenu = styled(DefaultSideMenu)`
